Fix the edit button for repos other than architecture (#66)
* Fix the edit button for repos other than architecture

I added some default runtime info for debugging. This got unintentionally
used in the actual GitHub Action. This broke the edit button for any repo
that didn't happen to match the default runtime, meaning I broke
everything but the architecture repo.

This change creates a new launch profile that we use for the actual GitHub
Action, allowing us to keep the easy debugging info.

I've changed the default info to reference a non-existant repo. This
prevents the "one repo works" problem, making it easier to identify when
things are broken.

* Update entrypoint.sh

* Update launchSettings.json