Fix action with global.json
Previously, if the repo using the action had a global.json at the root,
then he action's 'dotnet run' would use that. This would fail unless the
global.json happened to specify the same sdk version as the docker image
we're using.

With this change, we avoid this problem by explicitly running the action in
the src/Render directory.
avoid