Fix crash when template is null
Previously, we would check for a null template when deciding which template
to use, but not before doing a Path.Join with the null input. With this
change, we defer all use of the template input to when we're deciding which
template to use.

I don't love that Context is basically a God Object, but this is a small
step that gets a job done.